requirements.txt
python-telegram-bot==6.0.0
psycopg2
